Factor preflight class references out of site alias manager
